+++ swagger_parser/lib/src/config/swp_config.dart
@@ -28,8 +28,8 @@ class SWPConfig {
     this.replacementRules = const [],
     this.defaultContentType = 'application/json',
     this.extrasParameterByDefault = false,
+    this.dioOptionsParameterByDefault = false,
     this.pathMethodName = false,
-    this.requiredByDefault = true,
     this.mergeClients = false,
     this.enumsParentPrefix = true,
     this.skippedParameters = const <String>[],
@@ -55,8 +55,8 @@ class SWPConfig {
     required this.replacementRules,
     required this.defaultContentType,
     required this.extrasParameterByDefault,
+    required this.dioOptionsParameterByDefault,
     required this.pathMethodName,
-    required this.requiredByDefault,
     required this.mergeClients,
     required this.enumsParentPrefix,
     required this.skippedParameters,
@@ -123,6 +123,14 @@ class SWPConfig {
     if (extrasParameterByDefault is! bool?) {
       throw const ConfigException(
         "Config parameter 'extras_parameter_by_default' must be bool.",
+      );
+    }
+
+    final dioOptionsParameterByDefault =
+        yamlMap['dio_options_parameter_by_default'];
+    if (dioOptionsParameterByDefault is! bool?) {
+      throw const ConfigException(
+        "Config parameter 'dio_options_parameter_by_default' must be bool.",
       );
     }
 
@@ -287,7 +295,8 @@ class SWPConfig {
       defaultContentType: defaultContentType ?? dc.defaultContentType,
       extrasParameterByDefault:
           extrasParameterByDefault ?? dc.extrasParameterByDefault,
-      requiredByDefault: requiredByDefault ?? dc.requiredByDefault,
+      dioOptionsParameterByDefault:
+          dioOptionsParameterByDefault ?? dc.dioOptionsParameterByDefault,
       mergeClients: mergeClients ?? dc.mergeClients,
       enumsParentPrefix: enumsParentPrefix ?? dc.enumsParentPrefix,
       skippedParameters: skippedParameters ?? dc.skippedParameters,
@@ -370,23 +379,28 @@ class SWPConfig {
   /// Default content type for all requests and responses.
   ///
   /// If the content type does not match the default, generates:
-  /// @Headers(<String, String>{'Content-Type': 'PARSED CONTENT TYPE'})
+  /// `@Headers(<String, String>{'Content-Type': 'PARSED CONTENT TYPE'})`
   final String defaultContentType;
 
   /// DART ONLY
   /// Add extra parameter to all requests. Supported after retrofit 4.1.0.
   ///
   /// If  value is 'true', then the annotation will be added to all requests.
-  ///
+  /// ```dart
   /// @POST('/path/')
   /// Future<String> myMethod({@Extras() Map<String, dynamic>? extras});
+  /// ```
   final bool extrasParameterByDefault;
 
   /// DART ONLY
-  /// It is used if the value does not have the annotations `required` and `nullable`.
-  /// If the value is `true`, then value be `required`.
-  /// If the value is `false`, then `nullable`.
-  final bool requiredByDefault;
+  /// Add dio options parameter to all requests.
+  ///
+  /// If  value is 'true', then the annotation will be added to all requests.
+  /// ```dart
+  /// @POST('/path/')
+  /// Future<String> myMethod({@DioOptions() RequestOptions? options});
+  /// ```
+  final bool dioOptionsParameterByDefault;
 
   /// If `true`, use the endpoint path for the method name.
   /// if `false`, use `operationId`.
@@ -410,6 +424,7 @@ class SWPConfig {
       jsonSerializer: jsonSerializer,
       defaultContentType: defaultContentType,
       extrasParameterByDefault: extrasParameterByDefault,
+      dioOptionsParameterByDefault: dioOptionsParameterByDefault,
       rootClient: rootClient,
       rootClientName: rootClientName,
       clientPostfix: clientPostfix,
@@ -434,7 +449,6 @@ class SWPConfig {
       name: name,
       defaultContentType: defaultContentType,
       pathMethodName: pathMethodName,
-      requiredByDefault: requiredByDefault,
       mergeClients: mergeClients,
       enumsParentPrefix: enumsParentPrefix,
       skippedParameters: skippedParameters,