+++ swagger_parser/test/e2e/e2e_util.dart
@@ -30,12 +30,14 @@ Future<void> e2eTest(
     ),
   );
 
-  if (generateExpectedFiles) {
-    final directory = Directory(expectedFolderPath);
+  final expectedFilesFolder = Directory(expectedFolderPath);
+  if (generateExpectedFiles && expectedFilesFolder.existsSync()) {
+    expectedFilesFolder.deleteSync(recursive: true);
+  }
 
-    if (directory.existsSync()) {
-      directory.deleteSync(recursive: true);
-    }
+  final generatedFilesFolder = Directory(generatedFolderPath);
+  if (generatedFilesFolder.existsSync()) {
+    generatedFilesFolder.deleteSync(recursive: true);
   }
 
   await processor.generateFiles();